+++ sentry_kube/cli/run_job.py
@@ -75,8 +75,11 @@ def run_job(ctx, job_name, arg, kwarg, service_name, yes, quiet):
     if not (
         yes
         or click.confirm(
-            "Are you sure you want to apply this for customer "
-            f"{click.style(customer_name, fg='yellow', bold=True)}?"
+            "Are you sure you want to apply this for region "
+            f"{click.style(customer_name, fg='yellow', bold=True)}"
+            ", cluster "
+            f"{click.style(ctx.obj.cluster_name, fg='yellow', bold=True)}"
+            "?"
         )
     ):
         raise click.Abort()
@@ -119,6 +122,7 @@ def run_job(ctx, job_name, arg, kwarg, service_name, yes, quiet):
                         ctx.obj.cluster_name,
                         operation=f"run-job {job_name}",
                         service_name=service_name,
+                        quiet=ctx.obj.quiet_mode,
                     )
                 except Exception as e:
                     click.echo("!! Could not report an event to DataDog:")